# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * V1V Creative Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  password: string;
}

export interface LoginResponse {
  token: string;
  success: boolean;
}

export interface VerifyRequest {
  token: string;
}

export interface VerifyResponse {
  valid: boolean;
}

export interface Project {
  id: string;
  title: string;
  shortDesc: string;
  longDesc: string;
  images: string[];
  links: string[];
  tags: string[];
  year: string;
  category: string;
  featured: boolean;
}

export interface ProjectInput {
  title: string;
  shortDesc: string;
  longDesc: string;
  images?: string[];
  links?: string[];
  tags: string[];
  year: string;
  category: string;
  featured?: boolean;
}

export type SiteSettingsSocialLinks = {
  twitter?: string;
  instagram?: string;
  linkedin?: string;
  github?: string;
};

export type SiteSettingsThemeColors = {
  primary?: string;
  secondary?: string;
  accent?: string;
};

export interface SiteSettings {
  heroTagline: string;
  heroSubtitle: string;
  aboutTitle: string;
  aboutText: string;
  aboutFoundedYear?: string;
  contactEmail: string;
  contactPhone?: string;
  contactAddress?: string;
  socialLinks?: SiteSettingsSocialLinks;
  themeColors?: SiteSettingsThemeColors;
}