# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: V1V Creative Studio API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Admin authentication
  - name: projects
    description: Project management
  - name: settings
    description: Site settings management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid password

  /auth/verify:
    post:
      operationId: adminVerify
      tags: [auth]
      summary: Verify admin token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Token valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"

  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: Get all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get site settings
      responses:
        "200":
          description: Site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettings"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        success:
          type: boolean
      required:
        - token
        - success

    VerifyRequest:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    VerifyResponse:
      type: object
      properties:
        valid:
          type: boolean
      required:
        - valid

    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        shortDesc:
          type: string
        longDesc:
          type: string
        images:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        year:
          type: string
        category:
          type: string
        featured:
          type: boolean
      required:
        - id
        - title
        - shortDesc
        - longDesc
        - images
        - links
        - tags
        - year
        - category
        - featured

    ProjectInput:
      type: object
      properties:
        title:
          type: string
        shortDesc:
          type: string
        longDesc:
          type: string
        images:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        year:
          type: string
        category:
          type: string
        featured:
          type: boolean
      required:
        - title
        - shortDesc
        - longDesc
        - tags
        - year
        - category

    SiteSettings:
      type: object
      properties:
        heroTagline:
          type: string
        heroSubtitle:
          type: string
        aboutTitle:
          type: string
        aboutText:
          type: string
        aboutFoundedYear:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        contactAddress:
          type: string
        socialLinks:
          type: object
          properties:
            twitter:
              type: string
            instagram:
              type: string
            linkedin:
              type: string
            github:
              type: string
        themeColors:
          type: object
          properties:
            primary:
              type: string
            secondary:
              type: string
            accent:
              type: string
      required:
        - heroTagline
        - heroSubtitle
        - aboutTitle
        - aboutText
        - contactEmail