# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * V1V Creative Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  title: string;
  shortDesc: string;
  longDesc: string;
  images?: string[];
  links?: string[];
  tags: string[];
  year: string;
  category: string;
  featured?: boolean;
}