# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * V1V Creative Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { SiteSettingsSocialLinks } from "./siteSettingsSocialLinks";
import type { SiteSettingsThemeColors } from "./siteSettingsThemeColors";

export interface SiteSettings {
  heroTagline: string;
  heroSubtitle: string;
  aboutTitle: string;
  aboutText: string;
  aboutFoundedYear?: string;
  contactEmail: string;
  contactPhone?: string;
  contactAddress?: string;
  socialLinks?: SiteSettingsSocialLinks;
  themeColors?: SiteSettingsThemeColors;
}